I hate myself
Basically we're morons because we were drawing grid over grid and we
completely forgot what was taught in literally the last lecture.
Obviously the first grid was overriding the pixels drawn in second grid
because we were drawing everything on top of one another. Now I
JUST CHANGED THE SECOND THING TO BE DRAWN AND IT WORKED.